Initialize package-lock.json and update Next.js version in package.json
- Added a new `package-lock.json` file for dependency management.
- Updated the Next.js version from `15.3.1` to `^15.3.5` in `package.json` to ensure compatibility with the latest features and fixes.
- Introduced several new components for account settings, including views for changing email, phone number, and managing account deletion.
- Refactored the settings modal logic to improve state management and navigation between views.
