+++ apps/next/components/account/settings/AccountChangeEmailView.tsx
@@ -1,0 +1,5 @@
+import CreateAccountView from "./CreateAccountView";
+
+const AccountChangeEmailView = CreateAccountView("change", "email");
+
+export default AccountChangeEmailView;
