+++ apps/next/components/account/settings/AccountChangePhoneNumberView.tsx
@@ -1,0 +1,5 @@
+import CreateAccountView from "./CreateAccountView";
+
+const AccountChangePhoneNumberView = CreateAccountView("change", "phone-number");
+
+export default AccountChangePhoneNumberView;
