+++ apps/next/components/account/settings/AccountVerifyEmailView.tsx
@@ -1,0 +1,5 @@
+import CreateAccountView from "./CreateAccountView";
+
+const AccountVerifyEmailView = CreateAccountView("verify", "email");
+
+export default AccountVerifyEmailView;
