+++ apps/next/components/account/settings/AccountVerifyPhoneNumberView.tsx
@@ -1,0 +1,5 @@
+import CreateAccountView from "./CreateAccountView";
+
+const AccountVerifyPhoneNumberView = CreateAccountView("verify", "phone-number");
+
+export default AccountVerifyPhoneNumberView;
