+++ apps/next/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "next-app-template",
+  "name": "Konabra",
   "version": "0.0.1",
   "private": true,
   "scripts": {
@@ -57,7 +57,7 @@
     "lodash": "^4.17.21",
     "lucide-react": "^0.513.0",
     "mapbox-gl": "^3.12.0",
-    "next": "15.3.1",
+    "next": "^15.3.5",
     "next-themes": "0.4.6",
     "p-queue": "^8.1.0",
     "query-string": "^9.2.0",
